Refactors event subscription to use annotations and logging improvements
With this change you can set the logging level for the MoeCraft logs
to be different than the rest of Forge. By setting the environment
property "moecraft.logging.level" to the desired value.

This Change also configures the IDE run configurations to configure FML logging to
INFO and MoeCraft logging to DEBUG. So rerunning the generate runs
task will set you up.

Closes #19